# docs/swagger.yaml
basePath: /api/albums/feed
definitions:
  http.Error:
    properties:
      message:
        type: string
    type: object
...
  http.userInfoInput:
    properties:
      birthDate:
        $ref: '#/definitions/models.Date'
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      sex:
        $ref: '#/definitions/models.Sex'
    type: object
  http.userUploadAvatarResponse:
    properties:
      status:
        type: string
    type: object
...
  models.Date:
    properties:
      time.Time:
        type: string
    type: object
...
  models.Sex:
    enum:
    - M
    - F
    - O
    type: string
    x-enum-varnames:
    - Male
    - Female
    - Other
...
  models.UserTransfer:
    properties:
      avatarSrc:
        type: string
      birthDate:
        $ref: '#/definitions/models.Date'
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      sex:
        $ref: '#/definitions/models.Sex'
      username:
        type: string
    type: object
host: localhost:4443
info:
  contact:
    email: yarik1448kuzmin@gmail.com
    name: Fluire API Support
  description: Server API for Fluire Streaming Service Application
  title: Fluire API
  version: 1.0.1
paths:
...
  /api/users/{userID}/:
    get:
      description: Get user with chosen ID
      produces:
      - application/json
      responses:
        "200":
          description: User got
          schema:
            $ref: '#/definitions/models.UserTransfer'
        "400":
          description: Client error
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/http.Error'
        "403":
          description: Forbidden user
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Can't get user
          schema:
            $ref: '#/definitions/http.Error'
      summary: Get User
      tags:
      - User
  /api/users/{userID}/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Update user avatar
      parameters:
      - description: Avatar file
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Avatar updated
          schema:
            $ref: '#/definitions/http.userUploadAvatarResponse'
        "400":
          description: Invalid form data
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User Unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "403":
          description: User hasn't rights
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Upload Avatar
      tags:
      - User
...
  /api/users/{userID}/update:
    post:
      consumes:
      - application/json
      description: Update info about user
      parameters:
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/http.userInfoInput'
      produces:
      - application/json
      responses:
        "200":
          description: User info updated
          schema:
            $ref: '#/definitions/http.userUploadAvatarResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/http.Error'
        "401":
          description: User Unathorized
          schema:
            $ref: '#/definitions/http.Error'
        "403":
          description: User hasn't rights
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Can't change user info
          schema:
            $ref: '#/definitions/http.Error'
      summary: Update Info
      tags:
      - User
swagger: "2.0"